move gaia setup functions
