+++ tests/e2e/setup/gaia.go
@@ -3,6 +3,7 @@ package setup
 import (
 	"fmt"
 
+	"github.com/cosmos/cosmos-sdk/codec"
 	"github.com/cosmos/cosmos-sdk/crypto/keyring"
 )
 
@@ -14,4 +15,28 @@ type gaiaValidator struct {
 
 func (g *gaiaValidator) instanceName() string {
 	return fmt.Sprintf("gaiaval%d", g.index)
+}
+
+func (c *chain) createAndInitGaiaValidator(cdc codec.Codec) error {
+	// create gaia validator
+	gaiaVal := c.createGaiaValidator(0)
+
+	// create keys
+	mnemonic, info, err := createMemoryKey(cdc)
+	if err != nil {
+		return err
+	}
+
+	gaiaVal.keyInfo = *info
+	gaiaVal.mnemonic = mnemonic
+
+	c.GaiaValidators = append(c.GaiaValidators, gaiaVal)
+
+	return nil
+}
+
+func (c *chain) createGaiaValidator(index int) *gaiaValidator {
+	return &gaiaValidator{
+		index: index,
+	}
 }
